requirements.txt
astroid==2.2.5
Django==2.2.3
django-crispy-forms==1.7.2
django-pwa==1.0.5
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
Pillow==6.1.0
pylint==2.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
